+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/InMemoryFilmStorage.java
@@ -15,7 +15,7 @@ import java.util.Map;
 @Component
 @Slf4j
 public class InMemoryFilmStorage implements FilmStorage {
-    private final Map<Long, Film> films = new HashMap<>();
+    private final Map<Integer, Film> films = new HashMap<>();
     private int currentId = 0;
 
     @Override
@@ -27,7 +27,7 @@ public class InMemoryFilmStorage implements FilmStorage {
         }
         currentId++;
         film.setId((int) currentId);  // переназначаем id фильму
-        films.put((long) film.getId(), film);
+        films.put(film.getId(), film);
 
         log.debug("Фильм " + film.getName() + " создан.");
         return film;
@@ -39,7 +39,7 @@ public class InMemoryFilmStorage implements FilmStorage {
             log.debug("Фильм " + film.getName() + " не найден.");
             throw new NotFoundException("Такого фильма нет. Создайте новый.");
         }
-        films.put((long) film.getId(), film);
+        films.put(film.getId(), film);
         log.debug("Фильм " + film.getName() + " обновлен.");
         return film;
     }
@@ -60,7 +60,7 @@ public class InMemoryFilmStorage implements FilmStorage {
     }
 
     @Override
-    public Map<Long, Film> getFilms() {  // @FIXME: unused?
+    public Map<Integer, Film> getFilms() {  // @FIXME: unused?
         return films;
     }
